[DCP] Rewrite read slicing to use a wrapper.
Moved SlicedBufferedReader to utils and renamed to _ReaderView.

It no longer depends on file handles and is a pure wrapper. This
makes it general enought to handle non io stream objects like fsspec's.
